Move all router mocking logic into makeRouterObject
