+++ src/fetchData/fetchAppData.ts
@@ -1,6 +1,6 @@
 import { Fragment } from 'react';
 import type { AppContext, AppInitialProps } from 'next/app';
-import makeRouterObject from '../makeRouterObject';
+import makeRouterMock from '../makeRouterMock';
 import { makeGetInitialPropsContext } from './makeContextObject';
 import type {
   NextCustomAppFile,
@@ -19,13 +19,17 @@ export default async function fetchAppData({
 }): Promise<AppInitialProps | undefined> {
   const customApp = customAppFile.default;
   if (customApp.getInitialProps) {
+    const { asPath, pathname, query, route, basePath } = makeRouterMock({
+      pageObject,
+    });
+
     const ctx: AppContext = {
       // @NOTE AppTree is currently just a stub
       AppTree: Fragment,
       Component: pageObject.page.default,
       ctx: makeGetInitialPropsContext({ pageObject, options }),
       // @ts-ignore: Incomplete router object
-      router: makeRouterObject({ pageObject }),
+      router: { asPath, pathname, query, route, basePath },
     };
 
     const appInitialProps = await customApp.getInitialProps(ctx);
